feat(dotnet): use symbols to set conditions based on cron job or continuous job
