--- dotnet/iworker/Company.Worker1/WorkerCron.cs
@@ -1,30 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Company.Worker1
-{
-    public class Worker : BackgroundService
-    {
-        private readonly ILogger<Worker> _logger;
-        private readonly IHostApplicationLifetime _lifetime;
-
-        public Worker(ILogger<Worker> logger, IHostApplicationLifetime lifetime)
-        {
-            _logger = logger;
-            _lifetime = lifetime;
-        }
-
-        protected override async Task ExecuteAsync(CancellationToken stoppingToken)
-        {
-            _logger.LogInformation("Worker running at: {time}", DateTimeOffset.Now);
-
-            await Task.Delay(1000, stoppingToken);
-
-            _logger.LogInformation("Worker stopping at: {time}", DateTimeOffset.Now);
-            _lifetime.StopApplication();
-        }
-    }
-}
