--- dotnet/iworker/Company.Worker1/chart/valuesCron.yaml
@@ -1,28 +1,0 @@
-# Default values for chart.
-# This is a YAML-formatted file.
-# Declare variables to be passed into your templates.
-image:
-  repository: nginx
-  pullPolicy: IfNotPresent
-  tag: ""
-
-# Default Cron parameters.
-schedule: "* * * * *"
-restartPolicy: OnFailure
-concurrencyPolicy: Replace
-failedJobsHistoryLimit: 1
-successfulJobsHistoryLimit: 3
-
-nameOverride: ""
-fullnameOverride: ""
-
-podAnnotations: {}
-
-registry:
-  url: https://gitlab-registry.intility.com
-  user: ""
-  password: ""
-
-config: {}
-
-secrets: {}
